Split email service to interface and implementation
